moved to commands config package, created more beans to handle subscription and getting items
